solve clientside communication issue
game.ts now creates a room and puts all the sockets the game class
initialized with into that room. added a function in game.ts to add
event listeners to all the sockets in a game's room.

Now I'm running into a bug where it seems like the server is emitting
progressively more and more 'draw' events every time it receives a
correct guess from a client socket. it doesn't seem like any of the
emit events on clientside or serverside are firing >1 time, so it might
be something weird going on internally with socket.io. It also might
be something wrong with the way I'm passing functions between files.