fix >1 round bug i think???
the useEffect on TowerGame.tsx was removing the podium listener from the
socket, so it didn't have a way to exit the game. for some reason the
debugger in gameModes.ts saves the "this.gameState" from the previous game
only for the first guess, so it thinks a correct guess is wrong.

But it doesn't seem like an issue with the debugger - it fires
this.checkFilter(), then by some mystical force the guess loop re-runs
itself (I don't know how/why) and submits the correct guess successfully.

I want to investigate this further to make sure it's not affecting
the game or performance